refactor: color token control
