+++ packages/studio/src/controls/preset-options.ts
@@ -1,6 +1,6 @@
 import type { PresetBinding, ResolvedTokens } from "@sugarcube-sh/core/client";
 import type { PathIndex } from "../tokens/path-index";
-import { lastSegment, unwrapRef } from "./path-utils";
+import { lastSegment, resolveTerminalPath } from "./path-utils";
 
 export type ResolvedOption = {
     key: string;
@@ -41,16 +41,4 @@ function resolveGlobOptions(
             label: lastSegment(path),
             reference: `{${path}}`,
         }));
-}
-
-export function resolveTerminalPath(path: string, getToken: (path: string) => unknown): string {
-    const seen = new Set<string>();
-    let current = path;
-    while (true) {
-        if (seen.has(current)) return current;
-        seen.add(current);
-        const next = unwrapRef(getToken(current));
-        if (!next) return current;
-        current = next;
-    }
 }
